+++ api/web/src/components/CloudTAK/Menu/Mission/MissionInfo.vue
@@ -1,25 +1,49 @@
 <template>
 <div class='mx-2 my-2'>
-    <div class='datagrid'>
-        <div class='datagrid-item'>
+
+    <div class='mx-2 my-2'>
+        <div class='d-flex align-items-center'>
+            <button
+                v-if='subscribed === false'
+                class='btn btn-green'
+                style='height: 32px;'
+                @click='subscribe(true)'
+            >
+                Subscribe
+            </button>
+            <button
+                v-else-if='subscribed === true'
+                class='btn btn-danger'
+                style='height: 32px;'
+                @click='subscribe(false)'
+            >
+                Unsubscribe
+            </button>
+        </div>
+    </div>
+
+    <div class='row g-2'>
+        <div class='col-6'>
             <div class='datagrid-title'>
                 Created
             </div>
             <div
                 class='datagrid-content'
-                v-text='mission.createTime'
+                v-text='mission.createTime.replace(/T/, " ").replace(/:[0-9]+\..*/, "") + " UTC"'
             />
         </div>
-        <div class='datagrid-item'>
+        <div class='col-6'>
             <div class='datagrid-title'>
                 Subscribers
             </div>
+
+            <TablerLoading v-if='loading.users' :inline='true'/>
             <div
                 class='datagrid-content'
                 v-text='subscriptions.length'
             />
         </div>
-        <div class='datagrid-item'>
+        <div class='col-6'>
             <div class='datagrid-title'>
                 Contents
             </div>
@@ -28,7 +52,7 @@
                 v-text='Array.isArray(mission.content) ? mission.contents.length : 0 + " Items"'
             />
         </div>
-        <div class='datagrid-item'>
+        <div class='col-6'>
             <div class='datagrid-title'>
                 Groups (Channels)
             </div>
@@ -37,7 +61,7 @@
                 v-text='mission.groups.join(", ")'
             />
         </div>
-        <div class='datagrid-item'>
+        <div class='col-12'>
             <div class='datagrid-title'>
                 Description
             </div>
@@ -51,11 +75,62 @@
 </template>
 
 <script>
+import {
+    TablerLoading
+} from '@tak-ps/vue-tabler';
+
 export default {
     name: 'MissionInfo',
     props: {
         mission: Object,
-        subscriptions: Array
     },
+    components: {
+        TablerLoading
+    },
+    data: function() {
+        return {
+            loading: {
+                users: false
+            },
+            subscriptions: []
+        }
+    },
+    methods: {
+        fetchSubscriptions: async function() {
+            try {
+                const url = await stdurl(`/api/marti/missions/${this.mission.name}/subscriptions/roles`);
+                this.subscriptions = (await std(url)).data;
+            } catch (err) {
+                this.err = err;
+            }
+            this.loading.users = false;
+        },
+        subscribe: async function(subscribed) {
+            const layer = mapStore.getLayerByMode('mission', this.mission.guid);
+
+            if (subscribed === true && !layer) {
+                await mapStore.addDefaultLayer({
+                    id: this.mission.guid,
+                    url: `/mission/${encodeURIComponent(this.mission.name)}`,
+                    name: this.mission.name,
+                    source: this.mission.guid,
+                    type: 'geojson',
+                    before: 'CoT Icons',
+                    mode: 'mission',
+                    mode_id: this.mission.guid,
+                    clickable: [
+                        { id: `${this.mission.guid}-poly`, type: 'feat' },
+                        { id: `${this.mission.guid}-polyline`, type: 'feat' },
+                        { id: `${this.mission.guid}-line`, type: 'feat' },
+                        { id: this.mission.guid, type: 'feat' }
+                    ]
+                });
+            } else if (subscribed === false && layer) {
+                await mapStore.removeLayer(this.mission.name);
+            }
+
+            this.subscribed = !!mapStore.getLayerByMode('mission', this.mission.guid);
+        },
+    }
 }
 </script>
