+++ api/web/src/components/CloudTAK/Menu/Mission/MissionUsers.vue
@@ -1,25 +1,29 @@
 <template>
-    <div v-for='sub of subscriptions'>
-        <div class='col-12 py-2 px-2 d-flex hover-dark'>
-            <div class='row col-12 align-items-center'>
-                <div class='col-auto mx-2'>
-                    <div v-text='sub.username' />
-                    <div
-                        class='subheader'
-                        v-text='sub.username'
-                    />
-                </div>
-                <div class='col-auto ms-auto btn-list'>
-                    <IconUserBolt v-if='sub.role.type === "MISSION_OWNER"' size='32' v-tooltip='sub.role.type'/>
-                    <IconUserEdit v-else-if='sub.role.type === "MISSION_SUBSCRIBER"' size='32' v-tooltip='sub.role.type'/>
-                    <IconUser size='32' v-else-if='sub.role.type === "MISSION_READONLY_SUBSCRIBER"' v-tooltip='sub.role.type'/>
-                </div>
+<TablerNone v-if='!subscriptions.length' :create='false'/>
+<div v-else v-for='sub of subscriptions'>
+    <div class='col-12 py-2 px-2 d-flex hover-dark'>
+        <div class='row col-12 align-items-center'>
+            <div class='col-auto mx-2'>
+                <div v-text='sub.username' />
+                <div
+                    class='subheader'
+                    v-text='sub.username'
+                />
+            </div>
+            <div class='col-auto ms-auto btn-list'>
+                <IconUserBolt v-if='sub.role.type === "MISSION_OWNER"' size='32' v-tooltip='sub.role.type'/>
+                <IconUserEdit v-else-if='sub.role.type === "MISSION_SUBSCRIBER"' size='32' v-tooltip='sub.role.type'/>
+                <IconUser size='32' v-else-if='sub.role.type === "MISSION_READONLY_SUBSCRIBER"' v-tooltip='sub.role.type'/>
             </div>
         </div>
     </div>
+</div>
 </template>
 
 <script>
+import {
+    TablerNone
+} from '@tak-ps/vue-tabler';
 import {
     IconUserBolt,
     IconUserEdit,
@@ -28,13 +32,27 @@ import {
 
 export default {
     name: 'MissionUsers',
-    props: {
-        subscriptions: Array
-    },
     components: {
+        TablerNone,
         IconUserBolt,
         IconUserEdit,
         IconUser
+    },
+    data: function() {
+        return {
+            subscriptions: []
+        }
+    },
+    methods: {
+        fetchSubscriptions: async function() {
+            try {
+                const url = await stdurl(`/api/marti/missions/${this.mission.name}/subscriptions/roles`);
+                this.subscriptions = (await std(url)).data;
+            } catch (err) {
+                this.err = err;
+            }
+            this.loading.users = false;
+        },
     }
 }
 </script>
